+++ sdk/securityinsight/arm-securityinsight/src/securityInsights.ts
@@ -47,7 +47,9 @@ import {
   WatchlistItemsImpl,
   DataConnectorsImpl,
   DataConnectorsCheckRequirementsOperationsImpl,
-  OperationsImpl
+  OperationsImpl,
+  QueriesImpl,
+  EditedQueriesImpl
 } from "./operations";
 import {
   AlertRules,
@@ -82,7 +84,9 @@ import {
   WatchlistItems,
   DataConnectors,
   DataConnectorsCheckRequirementsOperations,
-  Operations
+  Operations,
+  Queries,
+  EditedQueries
 } from "./operationsInterfaces";
 import { SecurityInsightsOptionalParams } from "./models";
 
@@ -138,41 +142,34 @@ export class SecurityInsights extends coreClient.ServiceClient {
     };
     super(optionsWithDefaults);
 
-    let bearerTokenAuthenticationPolicyFound: boolean = false;
     if (options?.pipeline && options.pipeline.getOrderedPolicies().length > 0) {
       const pipelinePolicies: coreRestPipeline.PipelinePolicy[] = options.pipeline.getOrderedPolicies();
-      bearerTokenAuthenticationPolicyFound = pipelinePolicies.some(
+      const bearerTokenAuthenticationPolicyFound = pipelinePolicies.some(
         (pipelinePolicy) =>
           pipelinePolicy.name ===
           coreRestPipeline.bearerTokenAuthenticationPolicyName
       );
-    }
-    if (
-      !options ||
-      !options.pipeline ||
-      options.pipeline.getOrderedPolicies().length == 0 ||
-      !bearerTokenAuthenticationPolicyFound
-    ) {
-      this.pipeline.removePolicy({
-        name: coreRestPipeline.bearerTokenAuthenticationPolicyName
-      });
-      this.pipeline.addPolicy(
-        coreRestPipeline.bearerTokenAuthenticationPolicy({
-          credential: credentials,
-          scopes: `${optionsWithDefaults.credentialScopes}`,
-          challengeCallbacks: {
-            authorizeRequestOnChallenge:
-              coreClient.authorizeRequestOnClaimChallenge
-          }
-        })
-      );
+      if (!bearerTokenAuthenticationPolicyFound) {
+        this.pipeline.removePolicy({
+          name: coreRestPipeline.bearerTokenAuthenticationPolicyName
+        });
+        this.pipeline.addPolicy(
+          coreRestPipeline.bearerTokenAuthenticationPolicy({
+            scopes: `${optionsWithDefaults.baseUri}/.default`,
+            challengeCallbacks: {
+              authorizeRequestOnChallenge:
+                coreClient.authorizeRequestOnClaimChallenge
+            }
+          })
+        );
+      }
     }
     // Parameter assignments
     this.subscriptionId = subscriptionId;
 
     // Assigning values to Constant parameters
     this.$host = options.$host || "https://management.azure.com";
-    this.apiVersion = options.apiVersion || "2022-07-01-preview";
+    this.apiVersion = options.apiVersion || "2022-11-01-preview";
     this.alertRules = new AlertRulesImpl(this);
     this.actions = new ActionsImpl(this);
     this.alertRuleTemplates = new AlertRuleTemplatesImpl(this);
@@ -216,6 +213,8 @@ export class SecurityInsights extends coreClient.ServiceClient {
       this
     );
     this.operations = new OperationsImpl(this);
+    this.queries = new QueriesImpl(this);
+    this.editedQueries = new EditedQueriesImpl(this);
     this.addCustomApiVersionPolicy(options.apiVersion);
   }
 
@@ -280,4 +279,6 @@ export class SecurityInsights extends coreClient.ServiceClient {
   dataConnectors: DataConnectors;
   dataConnectorsCheckRequirementsOperations: DataConnectorsCheckRequirementsOperations;
   operations: Operations;
+  queries: Queries;
+  editedQueries: EditedQueries;
 }